40% ready

+++ app/users/models.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from sqlalchemy import BigInteger, Column, Integer, String
+from sqlalchemy import BigInteger, Column, Integer, String, BIGINT
 from sqlalchemy.orm import relationship
 
 from app.store.database.sqlalchemy_base import BaseModel
@@ -17,7 +17,7 @@ class UserModel(BaseModel):
     __tablename__ = "users"
 
     id = Column(BigInteger, primary_key=True)
-    tg_id = Column(Integer, nullable=False)
+    tg_id = Column(BIGINT, nullable=False)
     first_name = Column(String, nullable=False)
 
     game_scores = relationship("GameScoreModel", back_populates="player")
